add demo data for user and auth (#33)
* add demo data for user and auth

* fix lint

* fix lint fmt

+++ user/internal/models/user.go
@@ -5,27 +5,27 @@ import (
 )
 
 type User struct {
-	ID         uint      `gorm:"primary_key"`
+	ID         uint32    `gorm:"primary_key"`
 	Login      string    `gorm:"unique;not null"` // email
 	FullName   string    `gorm:"not null"`
-	TypeUserID uint      `gorm:"not null"`
+	TypeUserID uint32    `gorm:"not null"`
 	TypeUser   TypeUser  `gorm:"foreignKey:TypeUserID;constraint:OnUpdate:CASCADE,OnDelete:RESTRICT;"`
 	Status     bool      `gorm:"not null;default:true"`
 	Phone      string    `gorm:"not null"`
-	Rating     float64   `gorm:"default:0"`
+	Rating     float32   `gorm:"default:0"`
 	CreatedAt  time.Time `gorm:"autoCreateTime"`
 	UpdatedAt  time.Time `gorm:"autoCreateTime"`
 }
 
 // TypeUser customer or driver.
 type TypeUser struct {
-	ID   uint   `gorm:"primary_key"`
+	ID   uint32 `gorm:"primary_key"`
 	Type string `gorm:"not null"`
 }
 
 type Driver struct {
-	ID        uint   `gorm:"primaryKey"`
-	UserID    uint   `gorm:"not null"`
+	ID        uint32 `gorm:"primaryKey"`
+	UserID    uint32 `gorm:"not null"`
 	User      User   `gorm:"foreignKey:UserID;constraint:OnUpdate:CASCADE,OnDelete:RESTRICT;"`
 	License   string `gorm:"not null"`
 	CarModel  string `gorm:"not null"`
